Refactor folders in vmodel
