+++ app/src/main/java/com/blockeq/stellarwallet/activities/SendActivity.kt
@@ -25,8 +25,8 @@ import com.blockeq.stellarwallet.utils.AccountUtils
 import com.blockeq.stellarwallet.utils.NetworkUtils
 import com.blockeq.stellarwallet.utils.StringFormat.Companion.getNumDecimals
 import com.blockeq.stellarwallet.utils.StringFormat.Companion.hasDecimalPoint
-import com.blockeq.stellarwallet.vmodels.ExchangeEntity
-import com.blockeq.stellarwallet.vmodels.ExchangeViewModel
+import com.blockeq.stellarwallet.vmodels.exchange.ExchangeEntity
+import com.blockeq.stellarwallet.vmodels.exchange.ExchangeViewModel
 import com.davidmiguel.numberkeyboard.NumberKeyboardListener
 import kotlinx.android.synthetic.main.contents_send.*
