+++ app/src/main/java/com/blockeq/stellarwallet/vmodels/exchange/ExchangeProvider.kt
@@ -1,4 +1,4 @@
-package com.blockeq.stellarwallet.vmodels
+package com.blockeq.stellarwallet.vmodels.exchange
 
 import android.arch.persistence.room.Dao
 import android.arch.persistence.room.Insert
