+++ app/src/main/java/com/blockeq/stellarwallet/vmodels/exchange/ExchangesRoomDatabase.kt
@@ -1,4 +1,4 @@
-package com.blockeq.stellarwallet.vmodels
+package com.blockeq.stellarwallet.vmodels.exchange
 
 import android.arch.persistence.db.SupportSQLiteDatabase
 import android.arch.persistence.room.Database
