+++ app/src/main/java/com/blockeq/stellarwallet/vmodels/trading/TradingViewModel.kt
@@ -1,14 +1,11 @@
-package com.blockeq.stellarwallet.vmodels
+package com.blockeq.stellarwallet.vmodels.trading
 
 import android.app.Application
 import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.MutableLiveData
-import android.os.Handler
-import android.os.Looper
 import android.widget.Toast
 import com.blockeq.stellarwallet.models.*
 import com.blockeq.stellarwallet.remote.Horizon
-import org.stellar.sdk.requests.RequestBuilder
 import org.stellar.sdk.responses.OrderBookResponse
 import timber.log.Timber
 import java.util.*
